[#39] User API fetch me and show user's name on UI
