+++ nextjs-13/src/app/app.tsx
@@ -1,23 +1,56 @@
-"use client"
-import AuthLayout from "@components/AuthLayout";
-import AppLayout from "../components/AppLayout";
-import { usePathname } from 'next/navigation';
+"use client";
+
+import AppLayout from "@components/AppLayout";
+import {
+  UserState,
+  UserContext,
+  UserContextValue,
+} from '../context/user.context';
+import { useSession } from "next-auth/react"
+import { useCallback, useEffect, useMemo, useState } from 'react';
+import requestManager from '../lib/request/manager';
+// import 'stylesheets/application.scss';
+import { ApiMeResponse } from "./api/v1/me/route";
 
 export const metadata = {
   title: 'Next.js',
   description: 'Generated by Next.js',
 }
 
-export default function RootLayout({
+export default function App({
   children,
 }: {
   children: React.ReactNode
 }) {
-  const pathname = usePathname();
+  const { status } = useSession()
+  const [user, setUser] = useState<UserState>(undefined);
 
-  if (pathname.includes('/sign-in')) {
-    return <AuthLayout>{children}</AuthLayout>
-  }
+  const fetchCurrentUser = useCallback(async () => {
+    const { user: currentUser } = await requestManager<ApiMeResponse>(
+      'GET',
+      'v1/me'
+    );
 
-  return <AppLayout>{children}</AppLayout>
+    setUser(currentUser);
+  }, []);
+
+  useEffect(() => {
+    if (status === "authenticated") {
+      fetchCurrentUser();
+    }
+
+  }, [status, fetchCurrentUser]);
+
+  const userContextValue = useMemo<UserContextValue>(
+    () => ({ user, setUser }),
+    [user]
+  );
+
+  return (
+    <UserContext.Provider value={userContextValue}>
+      <AppLayout>
+          {children}
+      </AppLayout>
+    </UserContext.Provider>
+  )
 }
